Conflicts:
	ObjectAL.xcodeproj/project.pbxproj

+++ ObjectAL/ObjectAL/OALSimpleAudio.h
@@ -404,7 +404,7 @@ SYNTHESIZE_SINGLETON_FOR_CLASS_HEADER(OALSimpleAudio);
  */
 - (void) preloadEffects:(NSArray*) filePaths
 				   reduceToMono:(bool) reduceToMono
-		  progressBlock:(void (^)(uint progress, uint successCount, uint total)) progressBlock;
+		  progressBlock:(void (^)(NSUInteger progress, NSUInteger successCount, NSUInteger total)) progressBlock;
 
 #endif
